requirements.txt
Django==3.2
djangorestframework==3.14.0
PyJWT==2.6.0
psycopg2==2.8.6